requirements.txt
backoff=2.2.1
boto3=1.40.11
botocore=1.40.11
certifi=2025.8.3
charset-normalizer=3.4.3
faiss-cpu=1.12.0
filelock=3.19.1
fsspec=2025.7.0
hf-xet=1.1.7
huggingface-hub=0.34.4
idna=3.10
Jinja2=3.1.6
jmespath=1.0.1
joblib=1.5.1
json_repair=0.49.0
loguru=0.7.3
MarkupSafe=3.0.2
mpmath=1.3.0
networkx=3.4.2
numpy=2.2.6
nvidia-cublas-cu12=12.8.4.1
nvidia-cuda-cupti-cu12=12.8.90
nvidia-cuda-nvrtc-cu12=12.8.93
nvidia-cuda-runtime-cu12=12.8.90
nvidia-cudnn-cu12=9.10.2.21
nvidia-cufft-cu12=11.3.3.83
nvidia-cufile-cu12=1.13.1.3
nvidia-curand-cu12=10.3.9.90
nvidia-cusolver-cu12=11.7.3.90
nvidia-cusparse-cu12=12.5.8.93
nvidia-cusparselt-cu12=0.7.1
nvidia-nccl-cu12=2.27.3
nvidia-nvjitlink-cu12=12.8.93
nvidia-nvtx-cu12=12.8.90
packaging=25.0
pillow=11.3.0
pip=25.1
PyMuPDF=1.26.3
PyPDF2=3.0.1
python-dateutil=2.9.0.post0
python-dotenv=1.1.1
PyYAML=6.0.2
regex=2025.7.34
requests=2.32.4
s3transfer=0.13.1
safetensors=0.6.2
scikit-learn=1.7.1
scipy=1.15.3
sec-api=1.0.32
sentence-transformers=5.1.0
setuptools=78.1.1
six=1.17.0
sympy=1.14.0
threadpoolctl=3.6.0
tokenizers=0.21.4
torch=2.8.0
tqdm=4.67.1
transformers=4.55.2
triton=3.4.0
typing_extensions=4.14.1
urllib3=2.5.0
wheel=0.45.1